Factored naming related code out of diff and into new naming module.
git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@2726 0d8f46c8-9b30-49d0-b470-0f6283dc92e8
