+++ crds/svn_version.py
@@ -1,4 +1,4 @@
-__svn_version__ = '2723M'
+__svn_version__ = '2725M'
 
 __full_svn_info__ = '''
 Path: .
@@ -6,11 +6,11 @@ Working Copy Root Path: /Users/jmiller/work/workspace_crds/CRDS
 URL: https://aeon.stsci.edu/ssb/svn/crds/trunk
 Repository Root: https://aeon.stsci.edu/ssb/svn/crds
 Repository UUID: 0d8f46c8-9b30-49d0-b470-0f6283dc92e8
-Revision: 2723
+Revision: 2725
 Node Kind: directory
 Schedule: normal
 Last Changed Author: jmiller@stsci.edu
-Last Changed Rev: 2721
-Last Changed Date: 2015-11-05 17:48:19 -0500 (Thu, 05 Nov 2015)
+Last Changed Rev: 2725
+Last Changed Date: 2015-11-09 12:50:17 -0500 (Mon, 09 Nov 2015)
 
 '''
